requirements.txt
Flask>=2.2.2
requests>=2.28.1
openai-whisper>=20231117
torch>=2.0.0